% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsnz.R
\name{api_fetch}
\alias{api_fetch}
\title{Fetch from the NZ Statistics API}
\usage{
api_fetch(url_path, query = NULL)
}
\arguments{
\item{url_path}{The subpath of the base-URL.}

\item{query}{A query list for the API. Defaults to NULL (no query)}
}
\value{
The results of the GET request
}
\description{
Helper function to fetch information from the StatisticsNZ API
}

