% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsnz.R
\name{available_stats}
\alias{available_stats}
\title{available_stats}
\usage{
available_stats()
}
\value{
A data.frame with columns Stat and Description of statistics from the StatsNZ API.
}
\description{
Statistics available from the StatisticsNZ API
}

