% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsnz.R
\name{get_groups}
\alias{get_groups}
\title{get_groups}
\usage{
get_groups(stat)
}
\arguments{
\item{stat}{the statistic to query}
}
\value{
a vector of groups available for this statistic
}
\description{
Get the groups available for a particular statistic from the StatisticsNZ API
}

