% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsnz.R
\name{get_stats}
\alias{get_stats}
\title{Get data on a particular statistic}
\usage{
get_stats(stat, group = NULL)
}
\arguments{
\item{stat}{The statistic to fetch.}

\item{group}{The (optional) group of this statistic to fetch}
}
\value{
a data.frame containing the statistic of interest, usually a time-series
}
\description{
Get statistics data on a given statistic (optionally) and given group from the StatisticsNZ API
}

